# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Dough API
  version: "1.0.0"
  summary: A household's finances — accounts, ledger, budgets, investments and AI.
  description: |
    One versioned contract for every client: the web UI, and any native or
    third-party client that comes later.

    ## Response envelope

    Every response — success or failure — is one of two shapes.

    ```json
    { "success": true,  "data": ..., "meta": { ... } }
    { "success": false, "error": { "code", "message", "details" }, "meta": {...} }
    ```

    `meta.request_id` is the same value returned in the `X-Request-ID` header
    and written on every server log line for the request. Quote it in a bug
    report and the exact log lines can be found.

    Two documented exceptions:

    - `204 No Content` responses have no body at all. A body saying
      `{"success": true}` would be a 200 wearing the wrong status code.
    - The streaming endpoints answer `text/event-stream`. Failures decided
      *before* the stream opens are ordinary envelope errors; a failure
      mid-stream arrives as an SSE `error` event, which is the only channel left
      once the response has begun.

    ## Authentication

    Two credentials are accepted and they are for different callers.

    **API token** (`Authorization: Bearer dgh_...`) — for native and
    third-party clients. Obtain one from `POST /auth/login`. It is shown once
    and stored only as a hash, so it cannot be recovered; revoke and reissue if
...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /chat/conversations:
    get:
      tags: [chat]
      summary: Conversations, most recently updated first.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: sort
          in: query
          schema:
            type: string
            enum: [updated_at, created_at, title]
            default: updated_at
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: A page of conversations.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedSuccess'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Conversation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422': { $ref: '#/components/responses/ValidationError' }
    post:
      tags: [chat]
      summary: Start a conversation.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 80 }
      responses:
        '201':
          description: The new conversation.
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Conversation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /chat/conversations/{conversation_id}:
    get:
      tags: [chat]
      summary: One conversation.
      parameters: [{ $ref: '#/components/parameters/ConversationId' }]
      responses:
        '200':
          description: The conversation.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Conversation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [chat]
      summary: Rename a conversation.
      parameters: [{ $ref: '#/components/parameters/ConversationId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, maxLength: 80 }
      responses:
        '200':
          description: The renamed conversation.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Conversation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [chat]
      summary: Delete a conversation and its messages.
      parameters: [{ $ref: '#/components/parameters/ConversationId' }]
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /chat/conversations/{conversation_id}/messages:
    get:
      tags: [chat]
      summary: Messages in a conversation, oldest first.
      description: |
        Ordered by insertion, not by timestamp: both messages of a turn are
        often written in one commit and share a timestamp to the microsecond.
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        '200':
          description: The messages.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Message' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [chat]
      summary: Ask a question. Streams the reply.
      description: |
        Answers `text/event-stream`, not the envelope. Every refusal decidable
        before the stream opens — missing message, unknown conversation, no
        model configured — is an ordinary envelope error, so a client's normal
        error path covers everything except a failure mid-stream.

        Wire format:

        ```
        data: {"delta": "..."}     zero or more, in order
        data: {"error": "..."}     at most one, terminal
        data: [DONE]               on successful completion only
        ```

        `resend: true` regenerates the last answer without sending a new
        question, and requires the stored history to end with a user turn.
      parameters: [{ $ref: '#/components/parameters/ConversationId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message: { type: string, maxLength: 8000 }
                resend: { type: boolean, default: false }
                model: { type: string, maxLength: 80 }
      responses:
        '200':
          description: A server-sent event stream.
          content:
            text/event-stream:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    delete:
      tags: [chat]
      summary: Clear a conversation's messages, or truncate it.
      description: |
        `?keep=N` retains the first N messages instead of clearing everything,
        which is what the regenerate and edit-earlier-prompt actions need.
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - name: keep
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: How many messages survive.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          conversation_id: { type: string }
                          remaining: { type: integer }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------------------------------------------------------- copilot ----
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer dgh_...`. Obtain from `POST /auth/login`. Never
        send a token in a query string — the API does not read one, because
        query strings are written to access logs, browser history and `Referer`.
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        The web UI's credential. Unsafe methods additionally require
        `X-CSRF-Token`.
  parameters:
    Page:
      name: page
      in: query
      description: 1-based.
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      description: Clamped to the server maximum rather than refused.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Order:
      name: order
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
...
    ConversationId:
      name: conversation_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthenticated:
      description: No credential, or one that is not usable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: |
        Authenticated but not permitted. Read `error.code` to tell the three
        cases apart: `forbidden` (your role), `insufficient_scope` (your
        token's scopes), `csrf_failed` (a session request missing its token).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        No such resource — or one belonging to another household. The two are
        deliberately the same answer, so an id cannot be used as an oracle.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: |
        Well-formed and wrong. `error.details` names the offending fields.
        Distinct from 400, which means the request could not be parsed at all.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ServiceUnavailable:
      description: A dependency this endpoint needs is not configured or not answering.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Meta:
      type: object
      required: [api_version, timestamp]
      properties:
        api_version: { type: string, const: v1 }
        timestamp: { type: string, format: date-time }
        request_id:
          type: string
          description: Also returned as `X-Request-ID`, and on every server log line.
        pagination: { $ref: '#/components/schemas/Pagination' }
    Pagination:
      type: object
      required: [page, page_size, total, total_pages, has_next, has_prev]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
        has_next: { type: boolean }
        has_prev: { type: boolean }
    Success:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data: {}
        meta: { $ref: '#/components/schemas/Meta' }
    PagedSuccess:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            meta:
              allOf:
                - $ref: '#/components/schemas/Meta'
                - type: object
                  required: [pagination]
    Error:
      type: object
      required: [success, error, meta]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message:
              type: string
              description: For a person to read. May be reworded — never match on it.
            details:
              type: object
              description: Field-level breakdown. Present for `validation_error`.
        meta: { $ref: '#/components/schemas/Meta' }
    ErrorCode:
      type: string
      description: |
        The closed vocabulary a client should switch on.
      enum:
        - bad_request
        - validation_error
        - unauthenticated
        - forbidden
        - insufficient_scope
        - csrf_failed
        - not_found
        - conflict
        - rate_limited
        - service_unavailable
        - internal_error
        - method_not_allowed
        - payload_too_large
...
    Conversation:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Message:
      type: object
      properties:
        id: { type: integer }
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        created_at: { type: string, format: date-time }
...